The authentification was implemented

+++ blog/views.py
@@ -1,18 +1,22 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.views import View
-from .forms import AddBookForm
+
 from author.models import Author
+from .forms import AddBookForm
 from .models import Book
 
 
 class BookList(View):
     def get(self, request, book_id=-1):
-        books = Book.objects.all()
-        if book_id < 0:
-            context = {'object_list': books}
-            return render(request, "blog/index.html", context)
-        get_object_or_404(books, id=book_id).delete()
-        return redirect('home')
+        if request.user.is_authenticated:
+            books = Book.objects.all()
+            if book_id < 0:
+                context = {'object_list': books}
+                return render(request, "blog/index.html", context)
+
+            get_object_or_404(books, id=book_id).delete()
+            return redirect('home')
+        return render(request, "account/login.html")
 
 
 class BookDetail(View):
